Try to optimize timing around setting registers
This might be too aggressive. If clk_sys is 2.8 times as fast as clk_usb does that mean that I can set a register at clock time X and then set it again at X + 2 (ie - time 10 and time 12)? Or... does this mean that if I set X at time 10, I have to wait to time 13 to set it again?