Great a ZLP-free version of control transfer
Discovered that a set_address, while being an OUT for USB purposes, because it only sends back a ZLP/IN and the is "next" after the SETUP packet is sent, means that the ZLP/IN should make us specify "IN" for the USB controller. *Very tricky* to figure this out, but obvious now.